+++ proto/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'mimi.dart';
+import 'loginNavMain.dart';
 
 void main() {
   runApp(MyApp());
